SchemaGenerator - Simplify code-style for calling civicrm_navigation.tpl
There are a bunch of `*.tpl` files which are incorporated into the
baseline dataset. Most of these are listed in an array in `generateBasicData()`,
but `civicrm_navigation.tpl` was treated differently (with its own
`generateNavigation()`).

This cuts down several SLOC. It appears to be equivalent based on
reading and based on testing DB snapshots (before/after).