DrupalUtil::getDrupalSiteDir - Fix multisite. Use Drupal helper functions.
The previous code was copied over from `install/index.php`, which originally
worked in a Civi-only (no-CMS) environment.  It duplicated the D6/D7
`conf_path()` logic.  However, for `civicrm-setup`, we run everything in a
CMS-enabled environment, so it's fine to use Drupal helper functions like
`conf_path`.

This fixes multiple bugs with `getDrupalSiteDir()` - the net-effect of which
was that it produced the wrong values when using non-default multisites.